django-nonrel/ocl/wsgi.py
import os
import sys

path = '/opt/deploy/django/ocl'
if path not in sys.path:
    sys.path.append(path)

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'oclapi.settings.local')
os.environ.setdefault('DJANGO_CONFIGURATION', 'Local')

from configurations.wsgi import get_wsgi_application
application = get_wsgi_application()